+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 9.1.7 (2/26/2025 - )
 
 * sse: send ErrorResponse to client via "event: error" on exception
+* sse: log clientIP on sse:close action
 
 ### 9.1.6 (2/10/2025 - 2/25/2025)